+++ CHANGELOG.md
@@ -13,7 +13,8 @@
     - The record flow selector combo is now next to the tabs
     - The threads selector is now a menu button instead of a list view to gain some UI space
     - Automatically open first recorded thread
-    - Improved tutorial
+    - Improved graphical tutorial
+    - Help menu with tutorial and user's guide
     
 ### Bugs fixed